PXT-1228: typing info for video udfs that return pxt.Json
Also: finally reorganizes video.py into a separate submodule
